pyproject.toml
[project]
name = "security-kg"
version = "2.0.0"
description = "Convert security data (ATT&CK, CAPEC, CWE, CVE, CPE, D3FEND, ATLAS, CAR, ENGAGE, EPSS, KEV, Vulnrichment, GHSA, Sigma, ExploitDB) into SPO knowledge-graph triples (Parquet)"
requires-python = ">=3.11"
license = "Apache-2.0"
authors = [{ name = "Sougata Bhattacharya" }]
dependencies = [
    "mitreattack-python>=5.0.0",
    "pandas>=2.0.0",
    "pyarrow>=14.0.0",
    "pyyaml>=6.0.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "ruff>=0.4.0",
    "pre-commit>=3.0.0",
    "datasets>=2.14.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
]

[tool.ruff.lint.isort]
known-first-party = ["common", "convert", "convert_attack", "convert_atlas", "convert_capec", "convert_car", "convert_cpe", "convert_cve", "convert_cwe", "convert_d3fend", "convert_engage", "convert_epss", "convert_exploitdb", "convert_ghsa", "convert_kev", "convert_sigma", "convert_vulnrichment"]

[tool.pytest.ini_options]
markers = [
    "integration: tests that require network access (downloads real ATT&CK data)",
]
testpaths = ["tests"]
pythonpath = ["src"]
